pulling from main

+++ Blackjack/Player.java
@@ -3,25 +3,25 @@ package Blackjack;
 import java.util.ArrayList;
 
 public abstract class Player extends Person {
-    protected float bet;
+    protected int bet;
 
-    public Player(GameState game){
+    public Player(GameState game) {
         super(game);
         bet = 0;
     }
 
-    public Player(GameState game, float balance, ArrayList<Card> hand, float bet){
+    public Player(GameState game, int balance, ArrayList<Card> hand, int bet) {
         super(game, balance, hand);
         this.bet = bet;
     }
 
     //getter for bet
-    public float getBet(){
+    public int getBet() {
         return bet;
     }
 
     //setter for bet
-    public void setBet(float bet){
+    public void setBet(int bet) {
         this.bet = bet;
     }
 }
